Crate lmake_cargo_toml_to_md

lmake_cargo_toml_to_md

version: 0.5.4 date: 2020-07-15 authors: Luciano Bestia
Includes cargo.toml data into md.

Includes data from cargo.toml to md files.
To avoid out of sync data like version, authors and description.
The lmake_cargo_toml_to_md binary must be executed in the project root folder where is the cargo.toml file.
It works only for single projects and not for workspaces.

include cargo.toml data into md file

In the md file write these markers:

1 [comment]: # (lmake_cargo_toml_to_md start)
2 [comment]: # (lmake_cargo_toml_to_md end)

lmake_cargo_toml_to_md deletes the old lines between the markers
and includes the date and the cargo.toml data:
version, authors, description.

Tasks in Makefile.toml

I use cargo make to script the repetitive commands sequences.
https://github.com/sagiegurari/cargo-make
In Makefile.toml add a task like this:

[tasks.doc]
description = "create docs from doc comments"
clear = true
script = [
    "lmake_cargo_toml_to_md",
    "cargo doc --no-deps --document-private-items",
    "\\rsync -avz --delete-after target/doc/*  docs/",
]

Structs

REGEX_AUTHORS

REGEX_DESCRIPTION

REGEX_MD_END

REGEX_MD_START

REGEX_REMOVE_EMAIL

REGEX_VERSION

Functions

cargo_toml_to_md

utc_now
utc now